backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-mail==0.9.1
flask-cors==4.0.0
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
werkzeug==2.3.7 